implementação de rotate e meio de debugging

+++ include/Vec2.h
@@ -8,13 +8,16 @@
 #ifndef VEC2_H_
 #define VEC2_H_
 
+#include <cmath>
+#include <iostream>
+
 class Vec2 {
 public:
 	float x;
 	float y;
 	Vec2(float x, float y);
 	~Vec2();
-	Vec2 GetRotated( float ang );
+	Vec2 GetRotated( double ang );
 	Vec2 operator+(Vec2 v);
 };
